Add basic pattern matching
Functions can have several pattern / body pairs. The body corresponding
to the first matching pattern is evaluated.

These types of patterns can be matched now:

    1 => ... ; literal
    x => ... ; name
    (1, x) => ... ; list (literals and names can be mixed)

TODO: list (and vector and map, which don't exist yet) destructuring
patterns.

TODO: deconflict with currying. I think that the length of a function
will be defined as the length of the shortest pattern for the purpose of
currying.