Add multi-expression function bodies (#29)
Function bodies may now be one or more expressions. The return type is
the type of the last expression. The feature is designed to be like
`let` expressions, but more flexible (at the cost of purity) since
functions can now have side effects.

The  motivation for this is to allow `print` for debugging everywhere,
which I consider more important than purity for Peach.

Limitation: only the last expression may be tail
recursive for now.